The maintenance loot scarcity station trait now only spawns very rare loot  (#1500)
## About The Pull Request

The maintenance loot scarcity trait now only spawns very rare loot, with
only 1 instance being spawned.
This does not affect trash piles.

## Why It's Good For The Game

The station trait that removes maint loot is painful and boring to the
round because it means I can't powergame. This PR makes it so that the
trait isn't painful entirely.

## Proof Of Testing

If it compiles, it werks.

## Changelog

:cl: BurgerBB
balance: The maintenance loot scarcity station trait now only spawns
very rare loot
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>