Allows engineering cyborgs (and some saboteur cyborgs) to carry handheld tanks. (#1515)
## About The Pull Request

Adds a module that allows engineering cyborgs (and some saboteur
cyborgs) to carry handheld tanks.

Note that this does not apply to tank transfer valve bombs or single
tank bombs

## Why It's Good For The Game

Allows cyborgs to dick around with the RBMK and possibly other
unforeseen interactions. Anything bad a cyborg can do with a handheld
tank can already be done with a tank canister.

## Proof Of Testing

Untested.

## Changelog

:cl: BurgerBB
qol: Adds a module that allows engineering cyborgs (and some saboteur
cyborgs) to carry handheld tanks.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>