Fixes security jumpskirts (#1559)
## About The Pull Request
• Fixes jumpskirts being overidden by turning it into it's own type.
• Red jumpskirts and blue jumpskirts are now separate and both available
in the vendor.

## Why It's Good For The Game
No more bluesec bullshittery being forced on you.

## Proof Of Testing
errr arggh

## Changelog

:cl:
fix: Blue jumpskirts should not be in the vendor's red clothing section.
fix: Red jumpskirts should be available in the game again.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>