Fixes entombed reskin system (#1582)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
A fix for the entombed reskin system as provided by someone on discord.
Lets you reskin your entombed modsuit. Thats all the pull request does.
I am not versed enough in coding to understand what changes I made, but
from testing in visual studio code, it appears to work just fine with no
known issues.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
Not being able to reskin your modsuit was a bad thing because it was a
bug, this is a bugfix.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing

![image](https://github.com/Bubberstation/Bubberstation/assets/132009561/52488108-c33e-4ff6-83e1-ef30080f2853)

![image](https://github.com/Bubberstation/Bubberstation/assets/132009561/f784ae80-17d0-4cf6-a9af-9bc119d0cc90)

Incase those links didnt work, you can check it here. 
https://gyazo.com/d089b4505e5ab986237e5bb091685069
https://gyazo.com/2dfc0e928f9e78afea01e5641ed6f1a8
<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: Entombed trait reskin feature works now
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->