Storyteller port from horizon [DNM] (#456)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Ports storyteller from horizon, who deleted their repo. 

### How it works

_Written by Majkl-J_

Basically, the storyteller runs several "Tracks" that slowly fill up
with points. Upon reaching their set limit, an event is chosen and
spawned, and the storyteller takes a somewhat-random amount of points
from that track.


![image](https://github.com/Bubberstation/Bubberstation/assets/49160555/23b6f7ba-4d5a-45dc-9ed6-e88ed02ee019)

In this example, the moderate track will soon spawn an event.

Each track has its own sets of events it can spawn, sorted by intensity
(Roleset is midround and roundstart antags). Every event also has little
tags that the storyteller can be made to prioritize.

As with dynamic, events have a certain weight that determines the chance
they roll. Prioritized events get a multiplier to this weight.


![image](https://github.com/Bubberstation/Bubberstation/assets/49160555/136374ac-feb0-408f-89bf-c862167e2d47)

In this example, we can see some of the tags. Some storytellers can be
made to increase/decrease the weight of these events.

#### Extra stuff it does

Keeps track of med, engineering, and sec players. Whilst this still has
no actual usage, it is good to know it is capable of this.

## Why It's Good For The Game

Storyteller is an alternate event controller, in my opinion, superior to
dynamic.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: Nevimer, Majkl-J, BurgerBB, Azarak
add: Ported the storyteller event system from horizon
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: The Sharkening <95130227+StrangeWeirdKitten@users.noreply.github.com>
Co-authored-by: iero <>
Co-authored-by: Waterpig <wtryoutube@seznam.cz>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>
Co-authored-by: Return <donwest947@gmail.com>
Co-authored-by: BurgerLUA <8602857+BurgerLUA@users.noreply.github.com>