Adds a damage threshold + RNG requirement for EMPs affecting heads and torsos, as well as organs inside those two. (#1424)
## About The Pull Request

Adds a damage threshold + RNG requirement for EMPs affecting heads and
torsos, as well as organs inside those two.
For every point of damage on the bodypart (or if you're an organ, the
bodypart you belong to), there is a +2% chance for that bodypart (or
organ) to trigger the effects of EMP. This chance is doubled if it's a
heavy EMP.

Example, if your torso has 25 damage, and you were affected by an EMP,
your torso has a 50% chance to suffer the default effects of EMP. All
your organs inside your torso will EACH roll the 50% chance to be
affected by the EMP, regardless if your torso was EMP'd or not.

Graph here: https://www.desmos.com/calculator/7sdtn22uda

**Note that some effects can still trigger, such as stun and blindness,
regardless of damage. This is intentional.**

## Why It's Good For The Game

EMPs are super strong against synths. Perhaps unreasonably strong. This
PR nerfs this by making it so that you actually have to do some damage
to the synth before you can EMP them for an instant win. The
justification for this is below:


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/2b05fca7-df59-451d-a88e-eb59f51af478)

TL;DR, if your electronics are enclosed in steel, it won't suffer the
effects of EMP unless there are holes (damage).

## Proof Of Testing

Draft because untested.

## Frequently asked questions

### Synth main????
I have never played synth because I am not sexually attracted to metal.

### Does this affect cyborgs?
This PR does not change cyborg EMP behavior.

### Does this affect legs and arms?
This PR does not change leg and arm EMP behavior, since they're jointed.

## Changelog

:cl: BurgerBB
balance: Adds a damage threshold + RNG requirement for EMPs affecting
heads and torsos, as well as organs inside those two.
/:cl:

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>