[MIRROR] Tram malfunction adjustments (#28092)
* Tram malfunction adjustments (#83706)

## About The Pull Request

Adjusts tram malfunction and the tram malfunction event to be more
straightforward and similar to other telecoms machinery.

- Tram malfunction now requires engineering, borg, or AI to visit either
the tram or the tram controller in telecoms to press 'reset'
(previously: open the tram panel and find the appropriate tool to repair
it)
- Tram malfunction event only stops the tram once per event roll
- Adjust announcement to include message that engineers need to reset
the tram
- Removed loud priority_announce when the tram is fixed. Players know
the tram is fixed because they are using it again
- Tram doors no longer force crush during malfunction, since engineers
have no method to fix this
- EMP/communications blackout event will temporarily degrade crossing
signals, otherwise the tram continues to function normally
- Loss of telecoms power will degrade crossing signals, otherwise the
tram continues to function normally

Frequency of tram malfunction event is reduced to be in line with
communications blackout and grav gen blackout, and a maximum of 3
occurrences per round

Signal failure:
Previously: Signals fail until reset has physically been done on the
controller
New: Signal failure is moved to be part of the Communications Blackout
event, lasting the duration of EMP blast that hits telecoms

Communication loss:
Previously: For the duration of the Tram Malfunction event there is a 5%
chance the tram does a degraded stop, ending when the controller is
repaired or the timer ends
New: A guaranteed 1 degraded stop per event roll. To balance the
increase in occurrences it's been made easier to reset on the
controller. If the tram is not reset by the end of the timer, it is done
automatically

New telecoms room tram controller:

![dreamseeker_AGHpKvyunZ](https://github.com/tgstation/tgstation/assets/83487515/e8269607-11c9-407a-8f6e-d20f3f1b5f56)

## Why It's Good For The Game

Making the tram friendlier for engineering to pick up/understand, it
will function more predictably in the tram malfunction and
communications blackout event.

## Changelog

:cl: LT3
add: Telecoms now has a central tram controller
add: Tram controller is now included in communications blackout event
qol: Tram malfunction event only stops tram once, requiring engineering
to reset (no tools required)
qol: Tram malfunction event no longer sends a Central Command
announcement when fixed
del: Tram doors no longer force crush you during tram malfunction event
del: Silicons can no longer control the tram when communication is lost
/:cl:

* Tram malfunction adjustments

---------

Co-authored-by: lessthanthree <83487515+lessthnthree@users.noreply.github.com>
# Conflicts:
#	_maps/map_files/Birdshot/birdshot.dmm
#	_maps/map_files/tramstation/tramstation.dmm