Add Chief Screws Plush (#1676)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR adds a new plushie of one of our beloved Bubberbros,
Steals-The-Screwdriver. It adds a sprite-icon and object file to the
zubber modular
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
Personally I made this as an experiment with code, spriting, and BYOND
Dream Maker testing. Besides that, Screws is a popular character and
having a plushie of them is a wonderful Quality of life addition to
Bubber.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing

https://cdn.discordapp.com/attachments/734581723937112115/1254283661294047262/image.png?ex=6678ee30&is=66779cb0&hm=70df8bbcd69a0dd743ab94c7853cbb0c3b05c8fb2fc30b78b04554e6e2f5de9b&


https://cdn.discordapp.com/attachments/734581723937112115/1254283790919012362/image.png?ex=6678ee4f&is=66779ccf&hm=9b6ef39a084dda201fefa3de537655e1f879d15fd6b52d036e67dd0899934e31&


https://cdn.discordapp.com/attachments/734581723937112115/1254284198160764978/image.png?ex=6678eeb0&is=66779d30&hm=35e1b155c8842433382f7e7645f31fadd440ee970f5db2ef38a7bfc602372e8e&

NOTE: The sprite has been cleaned up a bit of the whiteness around its
horns since tested, and the 'bops' verb has been removed.

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: Rathen Ivanov

add: Added Employee Of The Month Special "Chief Screws Plush" for your
use and carrying!

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->