Bluespace Experi-Scanner (#1525)
## About The Pull Request

Adds a bluespace experi-scanner to the game which unlocks with the
Applied Bluespace tech node and can perform experiments instantly and
from range.

## Why It's Good For The Game

The first thirty minutes of science gameplay in a shift is typically the
exact same grindfest, and any roleplay from asking departments to open
for early upgrades/experiment scans loses all glamour after the first
few times a scientist performs it. The alternative to walking around
with a bluespace scanner is just building copies of the machines for
experiments in science anyway, which still lacks roleplay. QoL change
for science that encourages walking outside the department more without
counterhopping and allows it to engage in roleplay shticks sooner.

## Proof Of Testing

I ran and tested it locally.

## Changelog

:cl:
add: Added new bluespace experi-scanner.
image: Added a simple reskin of the experi-scanner for the bluespace
variant.
/:cl:

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>
Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>