Captain office access can't magically grant you AA anymore (#1615)
## About The Pull Request

See name, you can no longer give yourself ID modification access just by
having captain access

## Why It's Good For The Game

This is unintended behaviour, and results in certain roles abusing it to
grant themselves access without going through the respective heads of
staff.

This was already discussed in the team and with staff before. We
concluded it wasn't needed then and boiled it down to a specific player
issue. But since it keeps resurfacing repeatedly and shouldn't happen in
the first place, it is to be killed.

## Changelog

:cl:
fix: Captain office access can't magically grant you AA anymore
/:cl: