[Ready] Adds Big Bertha, a unique one-of-a-kind two-handed shield for security that blocks (almost) everything. (#1544)
## About The Pull Request

Adds Big Bertha, a unique one-of-a-kind two-handed shield for security
that has a base 100% chance to block. This is reduced by 1 per point of
armor penetration the damage deals.
Big Bertha spawns in the armory. **Only one exists.**
Big Bertha **slows you down when carried**, and **can only be carried
with two hands**. Has about the same offensive capabilities as a
toolbox.
When you block with the shield, it will deal half the damage blocked as
stamina damage to you, as long as the damage is above 10.
There is a chance to drop the shield when you block with it, equal to
how much damage to your stamina you have.

## Why It's Good For The Game

Security needs something for situations where there are only 2 officers
and there is a stealth box dsword hardsuit adrenals cqc stim meth
cocaine omnizine changeling heretic on the loose being super aggressive.

Since there is only one, it basically just means you shouldn't engage
this person as an antagonist with traditional weapons. This means that
there are diminishing returns with it when you have lets say 5 security
officers, as only one of them can actually carry it.

It's not good for chasing down antagonists since it does have a
slowdown. Note that it can't be affected by xenobiology speed potions
because Big Bertha is just so darn fat.

There is only one fear that I have with the weapon, and that is if a
holopara user gets the shield somehow and glues it to their hand. Admins
should deal with this by sending a deathsquad in.

## Proof Of Testing


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/9d75429b-1947-4c73-8515-82b6a2b2f093)
_Pictured: Warden out of brig._


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/48acc704-cdc6-4661-bb68-2942fa3f9373)
_Pictured: The shield doesn't make you immortal._

## Changelog

:cl: BurgerBB
add: Adds Big Bertha, a unique one-of-a-kind two-handed shield for
security that blocks literally everything. It spawns in the Armory.
/:cl:

---------

Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>