Reworks Sigma's Emitters (Again)
Ended up REALLY not being happy with this one. The emitter SMES is too hugboxy (to be more specific: it drains way too slowly to really be worth keeping; it was absurd how much wiggle room it afforded you) and caused an additional run_integration_tests issue in the process. Absolutely not. Not happening. Replaces the SMES with an additional pacman; a few sheets; and pretties up the wiring in the room a bit more. The five sheets should kill it being virtually indefinite.