More deployable pods! (#1695)
## About The Pull Request

I got caught in a rhythm ~~send me help, please, I can't stop making
pods, they're so neato...~~, I've made a bunch of pods which are
available via the NRI vendor, Izm-, Ish-, Izh- NRI guys... I can't
remember the name... There are a few variants, but most of them are
designed to be in sizes of 3, so they connect in a fairly modular
fashion.

Let's start at the top; 6x6, 3x3, Kitchen, Botany control, Botany trays,
O2 harvester, fan and comfort cabin.


![image](https://github.com/Bubberstation/Bubberstation/assets/53197594/a77a1777-a9fd-4039-b6f7-ef08288f0007)
The 6x6, Just primarily a large space for you to build in. 200cr


![image](https://github.com/Bubberstation/Bubberstation/assets/53197594/192d86b3-0ae5-405d-9e4a-cd48166a6de9)
The 3x3, A smaller space for building and maybe making corridors to
connect your pod base. 100cr.


![image](https://github.com/Bubberstation/Bubberstation/assets/53197594/cb01d7af-4803-4055-96e7-096b8200cfc0)
The kitchen! A useful little deployable, let's you really start to make
a home of it. 3x6. (the deploying tile will be between the oven and
condimaster) 400cr.


![image](https://github.com/Bubberstation/Bubberstation/assets/53197594/39909a62-8965-4dc3-b32d-629a0ed878a9)
The botany control capsule, this pod comes with a free tray capsule, as
well as a set of basic farming tools and the machines you'll need to get
yourself going. (Buy your own seeds though, we're not giving you
everything) 400cr


![image](https://github.com/Bubberstation/Bubberstation/assets/53197594/e356c35d-90af-4cec-8995-917bf88d0603)
Botany trays; This pod contains four trays. All basic, standard
unupgraded trays and a light... That's about it. 200cr.


![image](https://github.com/Bubberstation/Bubberstation/assets/53197594/f3be2604-db3d-4277-ad0a-db8cc1659132)
Here we have the O2 harvester. One of the only two pods with a fan to
secure the atmosphere. This will pull O2 out of the atmosphere, filter
it into a pump for your internals to be refilled from. 400cr.


![image](https://github.com/Bubberstation/Bubberstation/assets/53197594/0523daf0-1009-41f7-9df7-b70294bc951f)
The smallest of all of the pods. The Fan, this is designed to be an
airlock, place it outside of the doors if you want to keep the pressure
inside. Coming at a cost of 100cr each, they're not cheap, but not
expensive.


![image](https://github.com/Bubberstation/Bubberstation/assets/53197594/7326ac6a-2b96-4074-88f3-05a9d60b261a)
The cabin, a nice, cosy cabin designed for NRI officers made of cheap
wood but aesthetically pleasing enough. 200cr.

## Why It's Good For The Game

I wanted accessable pods for people besides miners or medical, you want
to have a nice little pod-glass base, you now can!

## Proof Of Testing

Screenshots in description.

## Changelog

:cl:
add: Added new pods!
/:cl:

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>