Adds Quote of the Round (#1687)
![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/80262905-7738-49a6-aa17-7b3845910f9d)


## About The Pull Request

Adds Quote of the Round, an entirely cosmetic way of showing off a
random ass message said on Common Radio. This does not ping people. By
default, it is posted in the same channel as roundstart announcements.

Between 1 and 2 hours (by default), a message will be chosen to be the
quote of the round. At the end of the round when the world reboots, the
message will be displayed.

It won't just choose any message. Sentence length is important (can't be
too long or too short), as well as the number of characters in the
message. This also can be configured via config options. This is in hope
that a funny message will be chosen and not someone just saying "Yeah."
on common, or "Science. research slut nanites."

## Why It's Good For The Game

It's funny.

## Proof Of Testing


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/a894b31e-97a3-4020-8da0-05edc95a8ada)

## Changelog

:cl: BurgerBB
add: Adds Quote of the Round, an entirely cosmetic way of showing off a
random ass message said on Common Radio. This does not ping people. By
default, it is posted in the same channel as roundstart announcements.
/:cl:

---------

Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>