Adds a GPS tracking collar (#1726)
## About The Pull Request
Adds a GPS tracking collar to the game. It can be found in the LustWish
premium section. The collar broadcasts a GPS signal with the name of the
collar. Tracking can be toggled on and off with alt-right-click. The
tracking collar can be locked, not only preventing removal but also
preventing toggling the tracking. The collar cannot receive GPS signals,
only send one.

## Why It's Good For The Game
Requested by Aeri on the Discord. Pre-approved by @projectkepler-ru. 

## Proof Of Testing
https://youtu.be/iFcZotyVUuQ

## Changelog

:cl:
add: Added a GPS tracking collar to the LustWish
/:cl:

---------

Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>