Adds a strict limit for how many antagonists of a type can be added in the round via storyteller (#1834)
## About The Pull Request

Adds a strict limit for how many antagonists of the same type can be
added in the round. This affects both midrounds and roundstarts.

### Current limits are:
- Bloodsucker: 3
- Changeling: 4
- Heretic: 2
- Malfunctioning AI: 1
- Nuke Ops: 5
- Spies: 4
- Traitor: 6

If an antagonist type is not listed here, there is no limit.

This does not affect an admin's ability to bus in additional antagonists
via the antag panel. This just affects storyteller.

## Why It's Good For The Game

Storyteller does not actually have a maximum cap of how many antagonists
of a certain type can be in the game. This means that while the maximum
heretics that storyteller can spawn roundstart is 2, there is actually
no limit (except for cost) for how many heretics can be in a round, as
the midround heretic event does not consider existing heretics in the
round.

This PR fixes that and adds a hard cap to antagonists of a certain type
in a round. This should prevent the rare and extremely stressful round
of 6 heretics spawning or something insane.

## Proof Of Testing

Untested. Can't really test this because I can't fake antag players.

## Changelog


:cl: BurgerBB
balance: Adds a strict limit for how many antagonists of the same type
can be added in the round. This affects both midrounds and roundstarts.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->