Revert "Nerfs No-Talk Borg Emags" (#1844)
Reverts Bubberstation/Bubberstation#1528

## About this PR

Reverts the 2 second emag time. I was kind of vibing with this at the
start, but I've really kind of changed my mind and I think this was a
sloppy way to handle this mechanic.

## Why it's good for the game

1. Emagged borgs have substantial counters where emagging one already
takes quite a bit of time (in the hall) for the average crewmember to
do. You need to still hack its cover, crowbar the borg, *then do a 2
second do_after* which makes it not really any feasable. It basically
railroads emagging borgs mostly to the roboticist now, which the
diffrence between an instant action and a !do_after isn't going to make
much of a diffrence.
2. Siliconnect litterally tells anyone who scans the borg who emagged
it. See a cover glitch? Use it.
3. It's not uncommon for an AI to immedietly notice that a borg was
emagged.
4. You can lock the borg down right after the fact. 
5. Cover glitch, obviously.
6. Flashes are not instant anymore.

Overall borgs have laws. Those laws can be changed. They can get hacked.
This is part of playing silicon and I really don't think it's healthy to
try balancing this away. But there's also litterally a "this person
hacked this borg" right click feature when you see a cover glitch.


🆑
balance: Reverted 2 second emag time on borgs
/:cl: