Moon Station Pharmafication (#1861)
## About The Pull Request
This PR renovates the Pharmacy and Chemistry Lab on Moon.

## Why It's Good For The Game
In its current state, both the lab and the pharmacy isn't comfortable to
use. There's a lot of wasted space that could be optimized. and chemists
shouldn't have to spend thirty minutes of a round to renovate the entire
thing. In the Chemistry Lab, moving the top table to the right frees up
a row, and machines are moved to the bottom left corner. Pharmacy
remains mostly the same, with reposition of items and machines. Grenade
casings, timer and and an igniter was missing as well. Some aesthetic
liberties were also taken.

## Proof Of Testing

![pharm](https://github.com/user-attachments/assets/b7b98806-2896-4ee7-b44d-2e967f7990a5)

![new](https://github.com/user-attachments/assets/2477ea1a-d23a-44f2-8908-f5f4bd1dce9e)

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
qol: Moon Station Pharmacy and Chemistry Lab is renovated.
/:cl: