Adding Haydeez (#1826)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
These changes add 8 new cyborg frames for 6 jobs and for syndicates.
This involved updating robot_defines, uploading the files, and adding
the robots to robot_model.dm
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
This was tested on a private server, and it all seemed to work properly.
the borgs do not have resting animations, but do have open cover
animations.
<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

[Sprite](https://github.com/SPLURT-Station/S.P.L.U.R.T-Station-13/pull/919)

## Why It's Good For The Game
This PR adds attractive borgs that multiple users from the server have
commented on positively about. While some may disparage the borg skins
for being too horny, it does not seem unrealistic to me that someone
would design a sexualized chassis for a being with no rights- in fact,
it seems inevitable. These are just reskins, so they do not add any new
mechanics outside of the visual
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing

![image](https://github.com/user-attachments/assets/54c80ece-e1b3-481d-8ccd-c9ac7b69b834)

![image](https://github.com/user-attachments/assets/8ea3093c-dfd7-4187-9109-7c11ac2cb808)

![image](https://github.com/user-attachments/assets/923359a6-da77-47fc-8325-3be8c351f74e)

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: Flvffi
added: 8 Haydee borg skins for Engineering, Medical, Mining, Janitorial,
Service, Peacekeeper, and Syndicate roles.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: The Sharkening <95130227+StrangeWeirdKitten@users.noreply.github.com>