Ethereal Cyborg Charging tweaks
Special thanks to Waterpig for help with this, this adds a new argument to the recharge station that allows you to set a limit on how much a thing is charged.  And adds an entry within the ethereal stomach that takes advantage of this new argument, allowing them to safely charge at cyborg chargers.  No more shall ethereals invoke the wrath of Zeus upon their compatriots for making the mistake of using a cyborg recharger!