Massive Overhaul of the Dauntless (Ready) (#1633)
## About The Pull Request

I just spent like the past 10 hours reworking the Dauntless map in a
post exam haze. Pretty much hit everything in some way or another
excluding the science area. I'm pretty much throwing this up than
heading off to pass out.


![image](https://github.com/Bubberstation/Bubberstation/assets/46614774/348ca2ae-4caa-4600-8cad-05da729379d0)

Bluespace Miners: Due to its current position being annoying to both the
kitchen and medical, it was moved to virology where it is only annoying
to medical. The current space does support additional miners which may
be an issue. Let me know if it is.

Engineering: The atmos tanks were shrunk to 2x1 with no space in between
the cells. Two additional tanks were added with gas miners for CO2 and
N2O. Additionally an electrolzyer was added. One additional SMES unit
was also added, however like the others it starts off totally
discharged. Lastly, if I did it right a 10 MJ capacity cell should be in
atmos. Powerflow monitoring computer moved to the window along with the
incinerator monitoring computer. Alert monitoring computer added.
Turbine space increased and input was moved to the north of the burn
chamber.

Medical : Removed the vending machine with grenade supplies as it was
redundant. Used space gained from removal of the miners to rework the
position of the surgical table most commonly used. Removed one of the
syndicate sleepers as they are redundant and moved some emergency
supplies over to where it was. **Added a genetics setup.** Let me know
if this is an issue as it was added in response to some issues with
humonkeys apparently not being good enough for dissection. Removed one
operating bay to replace virology. Removed one stasis bed. Added a limb
grower. Removed the crematorium as we have the option to toss bodies out
on lava if we really need to dump one and can't use the gibber. Dead
bodies added to morgue for flavor.

Brig: Holding cell replaced with a fully functional holding cell minus a
potential access issue I might still need to resolve. Cell comes with a
folding chair, toilet, and awful mattress and is strategically placed to
annoy the brig sergeant. Interrogation room upgraded with an intercomm
for listening in by 3rd parties. Kitchen provided to prisoners upgraded
with frontier equipment, including a flat pack drink dispenser and food
replicator. Smart Fridge set to food provided to alleviate lag related
to massive amounts of plants as it can happen. Bit running was moved to
where the MAA's office was. MAA's Office moved to old Liason's office.
Gave MAA's cat a toy.

Kitchen & Bar: Fridge moved to be south of the kitchen using part of the
space where the bluespace miners used to be. Additional space used to
increase size of kitchen to make it more useable. Smart fridge set for
food put in wall to bar. Chopping block, deep fryer, and food replicator
added to the kitchen. Coffee maker added to the bar along with the toy
mech display from the bridge. Butchering Knife and meathook added to the
freezer. One pacified syndicate mouseoperative added to eat any lose
crumbs.

Disposals: Lost a tile in height to support changes to engineering.

Cargo Bay: Added grating over cables and ducting to maintain OSHA
compliance. Additionally, per events giving the dauntless permission to
mine, added a mining slot and equipment.

Bridge: Added a comms monitoring setup.

Sauna : Given a pool and two bins of towels stacked on top of each
other.

Liason's office: Moved to be over by the crew area. Now features an
interdyne fax machine so they can represent an actual corporation.
**Massive buff given in the form of a fountain pin.** Provided an in
office coffee maker and carp skin chair.

Virology: Moved to medical with several tiles claimed from the lava to
make it fit. Miasma can moved to storage. Monkey provided replaced with
one that has human legs. I blame budget cuts.

Science : Added gas compressor and anomaly refinement. This should
provide partial coverage for mandatory experiments with further support
coming soon. Additionally slightly expanded xenobio and created a
separate circuits lab. Added an interdyne lathe.


![image](https://github.com/Bubberstation/Bubberstation/assets/46614774/c7ce48bf-391d-43d6-abaf-30d3ea2f678a)

Vault: Given a red circuit tile. **An emag was also placed in the safe
in case any of the preemagged machines the Dauntless gets are busted.**
Its also handy for exploring stuff like illegal fish in the portal/ Let
me know if this is problematic.

Dorms: Various posters have been put up. Dorms were given a second couch
for CAS or UNUM as we can have up to 8 crew members and 3 prisoners.
Each dorm room was given a plushie.

## Why It's Good For The Game

Beyond adding plushies to dorms, makes the ship in general far more
useable, adds the ability for detainment of rowdy crew and or prisoners
outside of the weird solitary cell in the back which is rather lacking,
and in general provides more space to roleplay.

## Proof Of Testing

Here is a pic from current engineering that I just snapped.


![image](https://github.com/Bubberstation/Bubberstation/assets/46614774/266865fd-a01f-4680-888d-fa70d57884b7)

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

:cl:

Add: Dauntless has been vastly reworked.
Add: The Interdyne Protolathe has been added. It is essentially a
reskinned Tarkon Lathe.
Add: Space Dauntless has been vastly reworked

/:cl:

---------

Co-authored-by: Swift <jackwars4@gmail.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>
Co-authored-by: The Sharkening <95130227+StrangeWeirdKitten@users.noreply.github.com>